fix(S1848): preserve reports for deferred DOM writes
